Return long instead of int when we are counting registration requests by activation key.
Should be in 76f5449418b5b68be1da31b3164258b914612a05 commit.

No functional changes.

+++ src/main/java/ru/mystamps/web/dao/UsersActivationDao.java
@@ -26,5 +26,5 @@ import ru.mystamps.web.entity.UsersActivation;
 
 public interface UsersActivationDao extends CrudRepository<UsersActivation, String> {
 	List<UsersActivation> findByCreatedAtLessThan(Date expiredSince);
-	int countByActivationKey(String activationKey);
+	long countByActivationKey(String activationKey);
 }
